major fix draft & worksheet
- add routes worksheet
- fix library commonlibs
- fix draft reviewer
- add status on worksheet and draft

+++ application/controllers/Draft.php
@@ -14,8 +14,9 @@ class Draft extends Operator_Controller
         $tot        = $this->draft->join('category')->join('theme')->orderBy('category.category_id')->orderBy('theme.theme_id')->orderBy('draft_id')->getAll();
 
         foreach ($drafts as $key => $value) {
-            $authors = $this->commonlibs->getIdAndName('author', 'draft_author', 'draft', $value->draft_id);
+            $authors = $this->draft->getIdAndName('author', 'draft_author', $value->draft_id);
             $value->author = $authors;
+            $value->status = $this->checkStatus($value->status);
         }
 
         $total     = count($tot);
@@ -44,7 +45,7 @@ class Draft extends Operator_Controller
             if ($upload) {
                 $input->draft_file =  "$draftFileName"; // Data for column "draft".
             }
-        }   
+        }
 
         if (!$this->draft->validate() || $this->form_validation->error_array()) {
             $pages     = $this->pages;
@@ -56,17 +57,37 @@ class Draft extends Operator_Controller
 
         $draft_id = $this->draft->insert($input);
 
+        $isSuccess = true;
+
         if ($draft_id > 0) {
             foreach ($input->author_id as $key => $value) {
-                    if ($this->commonlibs->insertIntoDifferentTable('draft_author', 'author', 'draft', $value, $draft_id)) {
-                    $this->session->set_flashdata('success', 'Data saved');
-                } else {
-                    $this->session->set_flashdata('error', 'Data author failed to save');
+                $data_author = array('author_id' => $value, 'draft_id' => $draft_id);
+
+                $draft_author_id = $this->draft->insert($data_author, 'draft_author');
+
+                if ($draft_author_id < 1 ) {
+                    $isSuccess = false;
+                    break;
                 }
             }
         } else {
-            $this->session->set_flashdata('error', 'Data failed to save');
+            $isSuccess = false;
         }
+
+        if ($isSuccess) {
+            $data_worksheet = array('draft_id' => $draft_id);
+            $worksheet_id = $this->draft->insert($data_worksheet, 'worksheet');
+
+            if ($worksheet_id < 1) {
+                $isSuccess = false;
+            }
+        }
+
+        if ($isSuccess) {
+            $this->session->set_flashdata('success', 'Data saved');
+        } else {
+            $this->session->set_flashdata('error', 'Data author failed to save');
+        } 
 
         redirect('draft');
       }
@@ -76,8 +97,15 @@ class Draft extends Operator_Controller
       public function edit($id = null)
 	{
         $draft = $this->draft->where('draft_id', $id)->get();
-        $author = $this->commonlibs->getIdAndName('author', 'draft_author', 'draft', $draft->draft_id);
-        $draft->author = $author;
+        $authors = $this->draft->getIdAndName('author', 'draft_author', $draft->draft_id);
+
+        $author_id = array();
+
+        foreach ($authors as $key => $value) {
+            $author_id[$key] = $value->author_id;
+        }
+
+        $draft->author_id = $author_id;
 
         if (!$draft) {
             $this->session->set_flashdata('warning', 'Draft data were not available');
@@ -85,12 +113,12 @@ class Draft extends Operator_Controller
         }
 
         if (!$_POST) {
-            $input = (object) $draft;$pages    = $this->pages;
+            $input = (object) $draft;
+            $pages    = $this->pages;
         } else {
             $input = (object) $this->input->post(null, true);
             $input->draft_file = $draft->draft_file; // Set draft file for preview.
         }
-
         
          if (!empty($_FILES) && $_FILES['draft_file']['size'] > 0) {
             // Upload new draft (if any)
@@ -117,24 +145,47 @@ class Draft extends Operator_Controller
             return;
         }
 
+        $isSuccess = true;
+
         if ($this->draft->where('draft_id', $id)->update($input)) {
             foreach ($input->author_id as $key => $value) {
-                if (isset($draft->author[$key])) {
-                    if ($this->commonlibs->updateIntoDifferentTable('draft_author', 'author', 'draft', $value, $id, $draft->author[$key]->author_id)) {
-                        $this->session->set_flashdata('success', 'Data updated');
+                if (isset($draft->author_id[$key])) {
+                    $draft_author_id = $this->draft->select('draft_author_id')
+                                                   ->where('author_id', $draft->author_id[$key])
+                                                   ->where('draft_id', $id)
+                                                   ->get('draft_author')
+                                                   ->draft_author_id;
+
+                    if ($draft_author_id > 0) {
+                        $isSuccess = true;
+                        $data_input = array('author_id' => $value);
+
+                        $this->draft->where('draft_author_id', $draft_author_id)
+                                    ->update($data_input, 'draft_author');
                     } else {
-                        $this->session->set_flashdata('error', 'Data failed to update');
+                        $isSuccess = false;
+                        break;
                     }
                 } else {
-                    if ($this->commonlibs->insertIntoDifferentTable('draft_author', 'author', 'draft', $value, $id)) {
-                        $this->session->set_flashdata('success', 'Data saved');
-                    } else {
-                        $this->session->set_flashdata('error', 'Data author failed to save');
+                    $isSuccess = true;
+                    $data_author = array('author_id' => $value, 'draft_id' => $id);
+
+                    $draft_author_id = $this->draft->insert($data_author, 'draft_author');
+
+                    if ($draft_author_id < 1) {
+                        $isSuccess = false;
+                        break;
                     }
                 }
             }
         } else {
-            $this->session->set_flashdata('error', 'Data failed to update');
+            $isSuccess = false;
+        }
+
+        if ($isSuccess) {
+            $this->session->set_flashdata('success', 'Data Updated');
+        } else {
+            $this->session->set_flashdata('error', 'Data Failed to Update');
         }
 
         redirect('draft');
@@ -149,12 +200,27 @@ class Draft extends Operator_Controller
             redirect('draft');
         }
 
-        if ($this->draft->where('draft_id', $id)->delete()) {
-            // Delete cover.
-            $this->draft->deleteDraftfile($draft->draft_file);
-            $this->commonlibs->deleteDifferentTable('draft_author', 'draft', $id);
+        $isSuccess = true;
+
+        $this->draft->where('draft_id', $id)
+                    ->delete('draft_author');
+
+        $affected_rows = $this->db->affected_rows();
+
+        if ($affected_rows > 0) {
+            if ($this->draft->where('draft_id', $id)->delete()) {
+                // Delete cover.
+                $this->draft->deleteDraftfile($draft->draft_file);
+            } else {
+                $isSuccess = false;
+            }
+        } else {
+            $isSuccess = false;
+        }
+
+        if ($isSuccess) {
             $this->session->set_flashdata('success', 'Data deleted');
-		} else {
+        } else {
             $this->session->set_flashdata('error', 'Data failed to delete');
         }
 
@@ -202,6 +268,69 @@ class Draft extends Operator_Controller
         $this->load->view('template', compact('pages', 'main_view', 'drafts', 'pagination', 'total'));
     }
 
+    public function checkStatus($code) {
+        $status = "";
+        switch ($code) {
+            case 0:
+                $status = 'Waiting for Worksheet';
+                break;
+            case 1:
+                $status = 'Choosing Reviewer';
+                break;
+            case 2:
+                $status = 'Reviewer Rejected';
+                break;
+            case 3:
+                $status = 'Process Review';
+                break;
+            case 4:
+                $status = 'Review Done';
+                break;
+            case 5:
+                $status = 'Choosing Editor';
+                break;
+            case 6:
+                $status = 'Process Editor';
+                break;
+            case 7:
+                $status = 'Review Done';
+                break;
+            case 8:
+                $status = 'Choosing Layouter';
+                break;
+            case 9:
+                $status = 'Process Layouter';
+                break;
+            case 10:
+                $status = 'Review Done';
+                break;
+            case 11:
+                $status = 'Choosing Proofread';
+                break;
+            case 12:
+                $status = 'Process Proofread';
+                break;
+            case 13:
+                $status = 'Review Done';
+                break;
+            case 14:
+                $status = 'Choosing Layouter';
+                break;
+            case 15:
+                $status = 'Process Layouter';
+                break;
+            case 16:
+                $status = 'Review Done';
+                break;
+            
+            default:
+                # code...
+                break;
+        }
+
+        return $status;
+    }
+
 
     /*
     |-----------------------------------------------------------------
@@ -226,7 +355,7 @@ class Draft extends Operator_Controller
         !$draft_id || $this->draft->where('draft_id !=', $draft_id);
         $draft = $this->draft->get();
 
-        if (count($draft)) {
+        if ($draft) {
             $this->form_validation->set_message('unique_draft_title', '%s has been used');
             return false;
         }
@@ -235,19 +364,19 @@ class Draft extends Operator_Controller
 
     public function unique_draft_title_author()
     {
-        $draft_title     = $this->input->post('draft_title');
-        $author_id     = $this->input->post('author_id');
-        $draft_id = $this->input->post('draft_id');
+        // $draft_title     = $this->input->post('draft_title');
+        // $author_id     = $this->input->post('author_id')[0];
+        // $draft_id = $this->input->post('draft_id');
         
-        $this->draft->where('author_id', $author_id);
-        $this->draft->where('draft_title', $draft_title);
-        !$draft_id || $this->draft->where('draft_id !=', $draft_id);
-        $draft = $this->draft->get();
+        // $this->draft->where('author_id', $author_id);
+        // $this->draft->where('draft_title', $draft_title);
+        // !$draft_id || $this->draft->where('draft_id !=', $draft_id);
+        // $draft = $this->draft->get();
 
-        if (count($draft)) {
-            $this->form_validation->set_message('unique_draft_title_author', 'Title and Author Name has been used');
-            return false;
-        }
+        // if ($draft) {
+        //     $this->form_validation->set_message('unique_draft_title_author', 'Title and Author Name has been used');
+        //     return false;
+        // }
         return true;
     }
